Build the harness without warnings, and keep it that way
Two warnings, both from GCC 11 on the runner and neither from clang.

cached_texture set texture_fullid::flag from a conditional whose arms
were df::texture_fullid_flag::mask_transparent_background and a bare 0,
one an enumeration and the other an int, which -Wextra objects to. Both
arms are now the bitfield's own uint32_t. The flag is one field being set
or cleared, so the two arms being the same type is what the code meant.

writerst::raw appended with vector::insert over a pair of pointers.
Reaching memcpy from there is a dozen inlined templates, and GCC 11 loses
the size of the destination on the way and reports a copy into a region
of size 0 when a cleared vector takes the file header. The warning is
wrong, the room is there, but resize and memcpy is the shorter path to
the same bytes and does not need the reader to know that. Measured: the
recording the codec test writes is byte-identical before and after, 407
bytes, same digest.

The workflow now passes -Werror through CXXFLAGS, which harness/build.sh
already honoured and harness/test.sh now does too. It is set in the
workflow rather than in the scripts so the build is held clean against
the one compiler the workflow pins, while a warning some other compiler
invents stays a warning on the machine that sees it instead of stopping
the work.

+++ smooth-movement.cpp
@@ -286,9 +286,12 @@ SDL_Texture *cached_texture(
 	texture_id.texpos=texpos;
 	texture_id.r=texture_id.g=texture_id.b=1.0f;
 	texture_id.br=texture_id.bg=texture_id.bb=0.0f;
+	// Both arms are the bitfield's own width. Writing the bare 0 instead makes one arm an
+	// enumeration and the other an int, which is a warning and reads as if the two were
+	// different kinds of thing; they are the same field, set and clear.
 	texture_id.flag=transparent_background?
-		df::texture_fullid_flag::mask_transparent_background:
-		0;
+		uint32_t(df::texture_fullid_flag::mask_transparent_background):
+		uint32_t(0);
 	const auto texture=renderer->tile_cache.tile_cache.find(texture_id);
 	return texture==renderer->tile_cache.tile_cache.end()?
 		nullptr:
